requirements.txt
python-dotenv>=0.19.1
azure-identity>=1.7.1
azure-keyvault-secrets>=4.3.0
